+++ nym-vpn-app/src/screens/index.ts
@@ -4,6 +4,8 @@ export { default as MainLayout } from './MainLayout';
 export { default as Error } from './Error';
 export { default as StartupError } from './StartupError';
 export { default as Welcome } from './Welcome';
+
+export * from './Onboarding';
 export * from './account';
 export * from './settings';
 export * from './node';